feat: spine通用
