+++ packages/plugin-renderer-spine38/lib/index.ts
@@ -1,3 +1,3 @@
 export { default as SpineSystem } from './SpineSystem';
 export { default as Spine } from './Spine';
-export type { SpineParams } from './Spine';
+export type { SpineParams } from '@eva/spine-base';
