+++ packages/spine-base/api-extractor.json
@@ -1,0 +1,7 @@
+{
+  "extends": "../../api-extractor.json",
+  "mainEntryPointFilePath": "./dist/packages/spine-base/lib/index.d.ts",
+  "dtsRollup": {
+    "publicTrimmedFilePath": "./dist/spine-base.d.ts"
+  }
+}
