+++ packages/spine-base/lib/TexCache.ts
